requirements.txt
Django==3.2.25
Pillow==9.5.0
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
python-dotenv==0.21.1
